+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### CHANGES
 - [Commands] Added EC2 commands: describe_ami, delete_ami, describe_snapshots, delete_snapshots
 - [Commands] Added S3 commands: delete_bucket
+- [Commands] Added SecretsManager commands: delete_secrets, list-secrets
 - [Dependencies] Upgraded boto3 to version 1.24.34
 - [Dependencies] Upgraded click to version 8.1.3